requirements.txt
PyGithub == 1.58.2